Add Linux (x64) support
Build:
- x64-linux-{debug,internal,release,trinitydev} presets chainloading the
  registry's x64-linux-carbon toolchain
- find_package(yaml): libyaml installs yamlConfig.cmake, which only matched
  YAML on case-insensitive filesystems
- the Objective-C++ sources are header-only on Linux
- Python tests set LD_LIBRARY_PATH to the Blue output directory: exefile
  dlopen()s blue by bare name, and unlike Windows and macOS the Linux loader
  does not fall back to the working directory
- vcpkg-registry submodule and baseline moved to the Linux triplets on the
  giobuilds fork

Portability fixes (same types on Windows/macOS, correct on LP64 Linux):
- BitPacker defines its uint64_t/int64_t overloads with those types instead
  of (unsigned) long long, and BlueNet packs uint64_t client ids
- BlueResFileSystemLocal includes <mutex> and returns an explicit empty
  vector where GCC read { {} } as std::in_place_t

Linux implementations:
- BlueSysInfo: XDG documents (user-dirs.dirs) and data directories (created
  on demand, as the spec asks), /proc/cpuinfo and /proc/meminfo, uname,
  /etc/machine-id, hostname split into machine and domain
- CLOCK_MONOTONIC performance timer, flock() in place of BSD O_SHLOCK and
  O_EXLOCK, newlocale/strto*_l for the YAML reader, GNU strerror_r
- errno-typed OsErrorType (glibc has no errno_t), OS::LINUX platform value,
  exefile pid scan over /proc, socket logger machine name and executable path
- error message boxes and clipboard report through stderr / failure until
  blue talks to a display server; OS language from LC_ALL/LC_MESSAGES/LANG

Fixes:
- BlueSetCrashReporter no longer installs the crash-reporter log echo when the
  host passes no reporter (exefile without Crashpad); every error log line
  dereferenced the null reporter
- TracksAllocatedAmount expects CCPMSize() of the block, which equals the
  request on Windows and macOS and is malloc_usable_size on glibc

389/389 tests pass natively on Fedora 44 and in the workspace build container.
The sysinfo and locale tests expect a desktop-like environment: a writable
$HOME, a machine-id, /usr/share/fonts and the fr_FR locale.

+++ include/BitPacker.h
@@ -587,7 +587,7 @@ bool BitPackerCore::Pack( const int64_t value )
 	if ( value == 0 )
 	{
 		QueueBits( (char*)&value, 1 );
-		Pack( 0ULL );
+		Pack( uint64_t( 0 ) );
 	}
 	else if ( value > 0 )
 	{
@@ -765,7 +765,7 @@ bool BitPackerCore::Pack( const double value )
 //------------------------------------------------------------------------------
 bool BitPackerCore::Pack( const double value, unsigned int places )
 {
-	unsigned long long val = 0;
+	uint64_t val = 0;
 	if ( value == 0 )
 	{
 		QueueBits( (char*)&val, 1 );
@@ -775,7 +775,7 @@ bool BitPackerCore::Pack( const double value, unsigned int places )
 	val = 1;
 	QueueBits( (char*)&val, 1 );
 
-	val = (unsigned long long)(value * (double)(1UL<<places));
+	val = (uint64_t)(value * (double)(1UL<<places));
 	return Pack( val );
 }
 
@@ -822,7 +822,7 @@ bool BitPackerCore::Unpack( int &value )
 }
 
 //------------------------------------------------------------------------------
-bool BitPackerCore::Unpack( unsigned long long &value )
+bool BitPackerCore::Unpack( uint64_t &value )
 {
 	value = 0;
 	DeQueueBits( (char*)&value, 3 );
@@ -836,18 +836,18 @@ bool BitPackerCore::Unpack( unsigned long long &value )
 }
 
 //------------------------------------------------------------------------------
-bool BitPackerCore::Unpack( long long &value )
+bool BitPackerCore::Unpack( int64_t &value )
 {
 	value = 0;
 	DeQueueBits( (char *)&value, 1 );
 	if ( value )
 	{
-		Unpack( (unsigned long long&)value );
+		Unpack( (uint64_t&)value );
 		value = -value;
 	}
 	else
 	{
-		Unpack( (unsigned long long&)value );
+		Unpack( (uint64_t&)value );
 	}
 
 	return Valid();
@@ -1020,7 +1020,7 @@ bool BitPackerCore::Unpack( double &value )
 //------------------------------------------------------------------------------
 bool BitPackerCore::Unpack( double &value, unsigned int places )
 {
-	unsigned long long val = 0;
+	uint64_t val = 0;
 	DeQueueBits( (char*)&val, 1 );
 
 	if ( val == 0 )
